requirements.txt
gradio
google-generativeai
python-dotenv
speechrecognition
pydub
gtts
elevenlabs
